Use correct replace_env_variable method.
