+++ tests/nlu/base/test_utils.py
@@ -111,7 +111,8 @@ def test_is_url():
 
 def test_environment_variable_not_existing():
     content = "model: \n  test: ${variable}"
-    with pytest.raises(KeyError):
+    with pytest.raises(ValueError
+                       ):
         rasa.utils.read_yaml(content)
